services/loaderStaging/requirements.txt
pymysql
pyyaml
python-dotenv


